Se agregan test de integración para las rutas de review

+++ movies-review-service/src/main/java/com/reactivespring/repository/MovieReviewRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
 import reactor.core.publisher.Flux;
 
 public interface MovieReviewRepository extends ReactiveMongoRepository<Review, String> {
-    Flux<Review> findReviewByMovieInfoId(Integer movieInfoId);
+    Flux<Review> findReviewByMovieInfoId(Long movieInfoId);
 }
